Poisson Reconstruction moved to header and an app is created which calls poisson reconstruction function and uses xml to read arguments
