+++ apps/PoissonRecon/CMakeLists.txt
@@ -2,14 +2,16 @@ cmake_minimum_required(VERSION 3.7)
 project(Poisson)
 
 set(PROJECT_SRCS
-${PROJECT_SOURCE_DIR}/src/PoissonRecon.cpp
+${PROJECT_SOURCE_DIR}/src/PoissonReconApp.cpp
 )
 
 source_group("src" FILES ${PROJECT_SRCS})
 
+find_package( tinyxml2 REQUIRED )
+
 add_executable(${PROJECT_NAME} ${PROJECT_SRCS})
 
-target_link_libraries(${PROJECT_NAME} PoissonRecon)
+target_link_libraries(${PROJECT_NAME} tinyxml2 PoissonRecon)
 
 set_property(TARGET ${PROJECT_NAME} PROPERTY FOLDER ${CMAKE_PROJECT_NAME}/apps)
