+++ apps/PoissonRecon/src/config.xml
@@ -1,0 +1,27 @@
+<CONFIG>
+	<INPUT>
+		<INPUT_PATH>/home/hypevr/PoissonRecon/clouds/Demo%03d.ply</INPUT_PATH>
+		<DEGREE>2</DEGREE>
+		<B_TYPE>3</B_TYPE>
+		<DEPTH>10</DEPTH>
+		<SCALE>1.1</SCALE>
+		<SAMPLES_PER_NODE>1.0</SAMPLES_PER_NODE>
+		<POINT_WEIGHT>4</POINT_WEIGHT>
+		<!--<CONFIDENCE></CONFIDENCE>-->
+		<!--<N_WEIGHTS></N_WEIGHTS>-->
+		<ITERS>8</ITERS>
+		<CG_DEPTH>0</CG_DEPTH>
+		<FULL_DEPTH>5</FULL_DEPTH>
+		<VOXEL_DEPTH>10</VOXEL_DEPTH>
+		<!--<PRIMAL_VOXEL></PRIMAL_VOXEL>-->
+		<COLOR>64</COLOR>
+		<DENSITY>--density</DENSITY>
+		<!--<LINEAR_FIT></LINEAR_FIT>-->
+		<!--<POLYGON_MESH></POLYGON_MESH>-->
+		<THREADS>6</THREADS>
+		<!--<VERBOSE></VERBOSE>-->
+	</INPUT>
+	<OUTPUT>
+		<OUTPUT_PATH>/home/hypevr/PoissonRecon/Demo%03d.ply</OUTPUT_PATH>
+	</OUTPUT>
+</CONFIG>
