+++ modules/PoissonRecon/CMakeLists.txt
@@ -7,6 +7,7 @@ ${PROJECT_SOURCE_DIR}/src/Factor.cpp
 ${PROJECT_SOURCE_DIR}/src/Geometry.cpp
 ${PROJECT_SOURCE_DIR}/src/MarchingCubes.cpp
 ${PROJECT_SOURCE_DIR}/src/PlyFile.cpp
+${PROJECT_SOURCE_DIR}/src/PoissonRecon.cpp
 )
 
 set(PROJECT_INCS
@@ -29,10 +30,12 @@ ${PROJECT_SOURCE_DIR}/include/Mesh/PoissonRecon/PointStream.h
 ${PROJECT_SOURCE_DIR}/include/Mesh/PoissonRecon/Polynomial.h
 ${PROJECT_SOURCE_DIR}/include/Mesh/PoissonRecon/PPolynomial.h
 ${PROJECT_SOURCE_DIR}/include/Mesh/PoissonRecon/SparseMatrix.h
+${PROJECT_SOURCE_DIR}/include/Mesh/PoissonRecon/PoissonRecon.h
 )
 
 source_group("include" FILES ${PROJECT_INCS})
 source_group("src"     FILES ${PROJECT_SRCS})
+
 
 add_library(${PROJECT_NAME} ${PROJECT_SRCS} ${PROJECT_INCS})
 
@@ -71,6 +74,7 @@ ${PROJECT_SOURCE_DIR}/include/Mesh/PoissonRecon/PointStream.h
 ${PROJECT_SOURCE_DIR}/include/Mesh/PoissonRecon/Polynomial.h
 ${PROJECT_SOURCE_DIR}/include/Mesh/PoissonRecon/PPolynomial.h
 ${PROJECT_SOURCE_DIR}/include/Mesh/PoissonRecon/SparseMatrix.h
+${PROJECT_SOURCE_DIR}/include/Mesh/PoissonRecon/PoissonRecon.h
         DESTINATION include/Mesh/PoissonRecon)
 
 export(TARGETS ${PROJECT_NAME}
